% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sd.R
\name{create_sd}
\alias{create_sd}
\title{Function to create an ODK Briefcase Storage directory}
\usage{
create_sd(path)
}
\arguments{
\item{path}{Directory path on which to create the ODK Briefcase Storage}
}
\value{
ODK Briefcase Storage directory in the specified path
}
\description{
Function to create an ODK Briefcase Storage directory
}
\examples{
create_sd(path = tempdir())

}
