% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_choice.R
\name{expand_choice}
\alias{expand_choice}
\title{Function to expand response to a more than one answer multiple choice
question coded as a concatenated string}
\usage{
expand_choice(
  df,
  x,
  values,
  pattern = "",
  prefix = x,
  labels = values,
  sep = "."
)
}
\arguments{
\item{df}{A dataframe containing the vector data that requires expansion.}

\item{x}{Name of variable in \code{df} containing vector needing expansion.}

\item{values}{Vector of string values used to create concatenated string
response.}

\item{pattern}{Pattern used to separate values in the concatenated string.
Default is "" for concatenated strings with no separator.}

\item{prefix}{Prefix to names of newly created variables.}

\item{labels}{Vector of names to use for columns of resulting data.frame.
If not specified, columns are named using \code{values}.}

\item{sep}{Charater to separate \code{prefix} from \code{labels} in the
names of newly created variables.}
}
\value{
A data.frame with same rows as \code{df} containing columns
  corresponding to each newly created variable.
}
\description{
Function to expand response to a more than one answer multiple choice
question coded as a concatenated string
}
\examples{
\dontrun{
expandMultipleChoice(df = individual,
                     x = "mddw1",
                     values = as.character(0:18),
                     pattern = " ",
                     prefix = "mddw",
                     sep = "")
}

}
