% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to get ODK survey datasets and forms from an ODK Aggregate server.}
\usage{
get_data(id, username, password, start = Sys.Date(), end = Sys.Date(),
  filename, overwrite = FALSE, exclude = TRUE, group.names = FALSE,
  split = FALSE, rep = FALSE, rep.name = NULL)
}
\arguments{
\item{id}{Form ID of form to get data from.}

\item{username}{ODK Aggregate server username credentials.}

\item{password}{ODK Aggregate server password credentials.}

\item{start}{Include data from submission dates after (inclusive) this
start date in export to CSV. Date format \code{<yyyy/MM/dd>} and default is
current system date}

\item{end}{Include data from submission dates before (exclusive) this date
in export to CSV. Date format is \code{<yyyy/MM/dd>} and default value is
current system date}

\item{filename}{Filename to use for data without the CSV file extension.}

\item{overwrite}{Overwrite existing output data in destination
directory with the same filename; default is FALSE}

\item{exclude}{Exclude media files on export; default is TRUE}

\item{group.names}{Logical. Should group names be removed from column names
on export? Default TRUE.}

\item{split}{Logical. Should select multiple fields be split on export?
Default FALSE.}

\item{rep}{Logical. Does the form have repeat/s? Default FALSE.}

\item{rep.name}{A vector of repeat names to read in the form. Default
is NULL. Must be specified if \code{rep} is TRUE.}
}
\value{
If \code{rep} is FALSE, a data.frame containing the dataset
  corresponding to the form ID specified. If \code{rep} is TRUE, a list of
  data.frames corresponding to the dataset from the main form of the survey
  and each of the repeats in the form.
}
\description{
This function is a wrapper for set of functions from the `odkr` package that
performs pulling of from from local or remote ODK directories and then
exports the embedded data as a comma-separate value (CSV) file.
}
\examples{
\dontrun{
  get_data(id = "stakholders",
           username = "validtrial",
           password = "zEF-STN-5ze-qom",
           filename = "test")
}

}
