% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkr.R
\docType{package}
\name{odkr}
\alias{odkr}
\alias{odkr-package}
\title{ODK Briefcase R API}
\description{
ODK Briefcase is a Java application for fetching and pushing OpenDataKit
(ODK) forms and their contents. It can be used to gather and aggregate data
from the mobile client ODK Collect when there is no internet connectivity,
when there is no ODK Aggregate Server set-up or when an ODK Aggregate
Server is not preferred. Current production version of ODK Briefcase is
v1.8.0 and an executable \code{.jar} file can be downloaded from the OpenDataKit
GitHub \href{https://github.com/opendatakit/briefcase/releases}{repo}.
}
\details{
As of ODK Briefcase version 1.4.4, a scriptable command line interface (CLI)
to the Java application has been available. This package provides an R
interface to ODK Briefcase via the available CLI to pull forms from a remote
ODK Aggregate Server or from a local ODK folder \code{/odk} collected from
mobile clients. This package requires Java 8 installed. Java 8 can be downloaded
\href{https://java.com/en/download/}{here}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/validmeasures/odkr}
  \item \url{https://validmeasures.org/odkr}
  \item Report bugs at \url{https://github.com/validmeasures/odkr/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernestgmd@gmail.com} (0000-0002-4887-4415) [copyright holder]

Authors:
\itemize{
  \item Laura Bramley \email{laura@validinternational.org} [copyright holder]
}

}
\keyword{internal}
