% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_local.R
\name{pull_local}
\alias{pull_local}
\title{pull_local}
\usage{
pull_local(briefcase = system.file("java",
  "ODK Briefcase v1.8.0 Production.jar", package = "odkr"), id, to = getwd(),
  from)
}
\arguments{
\item{briefcase}{Path to ODK Briefcase jar file; default is jar file
installed with package (ODK Briefcase v1.8.0)}

\item{id}{Form ID of form to be pulled}

\item{to}{Destination directory for pulled ODK forms; default destination
is at current working directory}

\item{from}{Source ODK directory (\code{/odk}) from ODK Collect mobile client}
}
\value{
Folder in destination directory named "ODK Briefcase Storage"
containing forms pulled from local ODK folder
}
\description{
Pull ODK forms from a local ODK folder (\code{/odk}) collected from
ODK Collect mobile clients
}
\examples{
  # Use pre-installed ODK Briefcase (version 1.8,0) and pull data from a
  # local ODK folder to current working directory
  pull_remote(id = "stakeholders",
              from = system.file("odk", package = "odkr"))

}
