% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkr.R
\docType{data}
\name{sampleData1}
\alias{sampleData1}
\title{Sample dataset from an impact evaluation study of a mother and child nutrition
programme in Kassala State, Sudan. This dataset contains cluster level data
from the survey.}
\format{
A data frame with 16 columns and 50 rows:
\describe{
\item{\code{admin.admin1.adm1}}{Date of survey}
\item{\code{admin.admin1.adm2}}{Enumerator ID}
\item{\code{admin.enameA}}{Name of enumerator (Arabic)}
\item{\code{admin.ename}}{Name of enumerator (English)}
\item{\code{admin.admin2.adm3}}{Survey round number}
\item{\code{admin.admin2.adm4}}{Study area / cluster}
\item{\code{location.loc1}}{Village ID}
\item{\code{location.loc1a}}{Village ID - other}
\item{\code{location.loc2}}{Is village a replacement village}
\item{\code{location.loc3}}{Village population}
\item{\code{location.loc4}}{GPS coordinates}
\item{\code{hh1}}{Household ID}
\item{\code{hh2}}{Number of women aged 15-49 in household}
\item{\code{wcount_count}}{Current woman respondent's ID}
\item{\code{KEY}}{Parent data identifier}
\item{\code{PARENT_KEY}}{Child data identifier}
}
}
\usage{
sampleData1
}
\description{
Sample dataset from an impact evaluation study of a mother and child nutrition
programme in Kassala State, Sudan. This dataset contains cluster level data
from the survey.
}
\keyword{datasets}
