% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_briefcase.R
\name{update_briefcase}
\alias{update_briefcase}
\title{update_briefcase}
\usage{
update_briefcase()
}
\description{
Updates pre-installed ODK Briefcase \code{jar} file to the latest version
downloaded from \url{https://opendatakit.org}.
}
