% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{export_data}
\usage{
export_data(briefcase = system.file("java", "odkBriefcase_v1.8.0.jar", package
  = "odkr"), id, from = getwd(), to = getwd(), filename = paste(id,
  "_data.csv", sep = ""), start = NULL, end = NULL, overwrite = FALSE,
  exclude = TRUE)
}
\arguments{
\item{briefcase}{Path to ODK Briefcase jar file; default is jar file
installed with package (ODK Briefcase v1.8.0)}

\item{id}{Form ID of form to be pulled}

\item{from}{Path to source ODK Briefcase Storage from which to extract data;
default is \code{getwd()}}

\item{to}{Destination directory to save output data file; default is
\code{getwd()}}

\item{filename}{Filename of output CSV data; default is
\code{paste(id, "_data.csv", sep = "")}}

\item{start}{Include data from submission dates after (inclusive) this start
date in export to CSV; format <yyyy/MM/dd>}

\item{end}{Include data from submission dates before (exclusive) this date in
export to CSV; format <yyyy/MM/dd>}

\item{overwrite}{Overwrite existing output data in destination
directory with the same filename; default is FALSE}

\item{exclude}{Exclude media files on export; default is TRUE}
}
\value{
CSV file in destination directory containing data from the pulled
forms
}
\description{
Export data in CSV format from local ODK Briefcase Storage directory to a
specified destination directory and a specified file name
}
\examples{
# Export data from a specified ODK Briefcase Storage directory to current
# working directory with a filename called "test.csv"
\dontrun{
  export_data(id = "stakeholders",
              filename = "test.csv",
              overwrite = TRUE)
}

}
