% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_briefcase.R
\name{get_briefcase}
\alias{get_briefcase}
\title{get_briefcase}
\usage{
get_briefcase(destination = system.file("java", package = "odkr"),
  filename = "odkBriefcase_latest")
}
\arguments{
\item{destination}{Path to directory where ODK Briefcase \code{.jar} file
will be downloaded into. Default is the \code{java} directory in the
\code{odkr} package installation}

\item{filename}{Filename of the downloaded ODK Briefcase \code{.jar} file.
Default is \code{"odkBriefcase_latest"}}
}
\description{
Updates pre-installed ODK Briefcase \code{jar} file to the latest version
downloaded from \url{https://opendatakit.org}.
}
\examples{
  #

}
