% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_help.R
\name{get_help}
\alias{get_help}
\title{get_help}
\usage{
get_help(briefcase = system.file("java", "odkBriefcase_v1.8.0.jar", package =
  "odkr"))
}
\arguments{
\item{briefcase}{Path to ODK Briefcase jar file; default is jar file
installed with package (ODK Briefcase v1.8.0)}
}
\value{
Help notes on usage of ODK Briefcase via command line interface
}
\description{
Get help with command line interface for ODK Briefcase
}
\examples{
get_help()

}
