% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_help.R
\name{get_help}
\alias{get_help}
\title{get_help}
\usage{
get_help(briefcase = system.file("java", "odkBriefcase_latest.jar", package =
  "odkr"))
}
\arguments{
\item{briefcase}{Path to pre-installed ODK Briefcase \code{jar} file. Default
is \code{jar} file (currently ODK Briefcase v1.8.0) installed with the
current version of the package accessed via \code{system.file()}. Path
can be set to access an ODK Briefcase \code{jar} file downloaded locally
in user's machine}
}
\value{
Help notes on usage of ODK Briefcase via command line interface
}
\description{
Get help with command line interface for ODK Briefcase
}
\examples{
get_help()

}
