% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_local.R
\name{pull_local}
\alias{pull_local}
\title{pull_local}
\usage{
pull_local(briefcase = system.file("java", "odkBriefcase_latest.jar", package
  = "odkr"), id, to = getwd(), from, pem = NULL)
}
\arguments{
\item{briefcase}{Path to pre-installed ODK Briefcase \code{jar} file. Default
is \code{jar} file (currently ODK Briefcase v1.8.0) installed with the
current version of the package accessed via \code{system.file()}. Path
can be set to access an ODK Briefcase \code{jar} file downloaded locally
in user's machine}

\item{id}{Form ID of form to be pulled}

\item{to}{Destination directory for pulled ODK forms; default destination
is at current working directory}

\item{from}{Source ODK directory (\code{/odk}) from ODK Collect mobile client}

\item{pem}{If form to be pulled is encrypted, a PEM private key file would be
required to pull forms; default is NULL; if form is encrypted, provide path
to PEM file}
}
\value{
Folder in destination directory named "ODK Briefcase Storage"
containing forms pulled from local ODK folder
}
\description{
Pull ODK forms from a local ODK folder (\code{/odk}) collected from
ODK Collect mobile clients
}
\examples{
# Use pre-installed ODK Briefcase (version 1.8.0) and pull forms from a
# local ODK folder found in Desktop to current working directory
\dontrun{
  pull_local(id = "stakeholders",
             from = ~/Desktop/odk)
}

}
