% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_remote.R
\name{pull_remote}
\alias{pull_remote}
\title{pull_remote}
\usage{
pull_remote(briefcase = system.file("java", "odkBriefcase_latest.jar", package
  = "odkr"), id, to = getwd(), from, username, password)
}
\arguments{
\item{briefcase}{Path to pre-installed ODK Briefcase \code{jar} file. Default
is \code{jar} file (currently ODK Briefcase v1.8.0) installed with the
current version of the package accessed via \code{system.file()}. Path
can be set to access an ODK Briefcase \code{jar} file downloaded locally
in user's machine}

\item{id}{Form ID of form to be pulled}

\item{to}{Destination directory for pulled ODK forms; default destination
is at current working directory}

\item{from}{URL of remote ODK Aggregate server}

\item{username}{Username for account in remote ODK Aggregate server from
which forms are to be pulled}

\item{password}{Password for account in remote ODK Aggregate server from
which forms are to be pulled}
}
\value{
Folder in destination directory named "ODK Briefcase Storage"
containing forms pulled from remote ODK Aggregate server
}
\description{
Pull ODK forms from remote ODK Aggregate via ODK Briefcase
}
\examples{
  # Use pre-installed ODK Briefcase (version 1.8,0) and connect to a test
  # remote ODK Aggregate server from ONA (https://ona.io); pulled forms to
  # be saved in default location at current working directory
  pull_remote(id = "stakeholders",
              from = "https://ona.io/validtrial",
              username = "validtrial",
              password = "zEF-STN-5ze-qom")

}
