% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranameODK.R
\name{renameODK}
\alias{renameODK}
\title{ranameODK}
\usage{
renameODK(data)
}
\arguments{
\item{data}{Dataframe object of dataset exported from ODK Aggregate Server}
}
\value{
Data frame object with renamed variables
}
\description{
Rename column names of data exported from an ODK Aggregate Server or from
ODK Briefcase into more usable and human readable variable names.
}
